+++ src/node/node_state.h
@@ -306,15 +306,17 @@ namespace ccf
     QuoteVerificationResult verify_quote(
       kv::ReadOnlyTx& tx,
       const QuoteInfo& quote_info,
-      const std::vector<uint8_t>& expected_node_public_key_der) override
+      const std::vector<uint8_t>& expected_node_public_key_der,
+      CodeDigest& code_digest) override
     {
 #ifdef GET_QUOTE
       return enclave_attestation_provider.verify_quote_against_store(
-        tx, quote_info, expected_node_public_key_der);
+        tx, quote_info, expected_node_public_key_der, code_digest);
 #else
       (void)tx;
       (void)quote_info;
       (void)expected_node_public_key_der;
+      (void)code_digest;
       return QuoteVerificationResult::Verified;
 #endif
     }
@@ -358,7 +360,15 @@ namespace ccf
 #ifdef GET_QUOTE
       quote_info = enclave_attestation_provider.generate_quote(
         node_sign_kp->public_key_der());
-      node_code_id = enclave_attestation_provider.get_code_id(quote_info);
+      auto code_id = enclave_attestation_provider.get_code_id(quote_info);
+      if (code_id.has_value())
+      {
+        node_code_id = code_id.value();
+      }
+      else
+      {
+        throw std::logic_error("Failed to extract code id from quote");
+      }
 #endif
 
       switch (start_type)
@@ -990,6 +1000,19 @@ namespace ccf
         self = NodeId(fmt::format("{:#064}", id.value()));
       }
 
+      CodeDigest code_digest;
+#ifdef GET_QUOTE
+      auto code_id = enclave_attestation_provider.get_code_id(quote_info);
+      if (code_id.has_value())
+      {
+        code_digest = code_id.value();
+      }
+      else
+      {
+        throw std::logic_error("Failed to extract code id from quote");
+      }
+#endif
+
       g.add_node(
         self,
         {node_info_network,
@@ -997,7 +1020,9 @@ namespace ccf
          quote_info,
          node_encrypt_kp->public_key_pem().raw(),
          NodeStatus::PENDING,
-         get_next_reconfiguration_id(network, tx)});
+         get_next_reconfiguration_id(network, tx),
+         std::nullopt,
+         ds::to_hex(code_digest.data)});
 
       LOG_INFO_FMT("Deleted previous nodes and added self as {}", self);