+++ tests/nobuiltins.py
@@ -8,7 +8,7 @@ import openapi_spec_validator
 
 
 def test_nobuiltins_endpoints(network, args):
-    primary, _ = network.find_primary()
+    primary, backups = network.find_nodes()
     with primary.client() as c:
         r = c.get("/app/commit")
         assert r.status_code == HTTPStatus.OK
@@ -32,6 +32,16 @@ def test_nobuiltins_endpoints(network, args):
         body_j = r.body.json()
         assert body_j["transaction_id"] == f"{tx_id}"
 
+        r = c.get("/app/all_nodes")
+        assert r.status_code == HTTPStatus.OK
+        body_j = r.body.json()
+        known_node_ids = [node.node_id for node in (primary, *backups)]
+        for node_id, node_info in body_j["nodes"].items():
+            assert (
+                node_id in known_node_ids
+            ), f"Response contains '{node_id}', which is not in known IDs: {known_node_ids}"
+            assert node_info["quote_format"] == "OE_SGX_v1"
+
 
 def run(args):
     with infra.network.network(
@@ -45,5 +55,5 @@ def run(args):
 if __name__ == "__main__":
     args = infra.e2e_args.cli_args()
 
-    args.nodes = infra.e2e_args.min_nodes(args, f=0)
+    args.nodes = infra.e2e_args.min_nodes(args, f=1)
     run(args)